+++ sdk/textanalytics/azure-ai-textanalytics/CHANGELOG.md
@@ -5,6 +5,12 @@
 - We are now targeting the service's v3.1 API as the default instead of v3.1-preview.4.
 - Added a new class, `HealthcareEntityCategory` to replace the `String` type of property `category` in the `HealthcareEntity`.
 
+### Breaking Changes
+- Removed `StringIndexType`, the type will be shared in the Azure Core.
+- Renamed type `PiiEntityDomainType` to `PiiEntityDomain`.
+- Renamed property name `result` to `documentsResults` in `AnalyzeSentimentActionResult`, `ExtractKeyPhrasesActionResult`,
+  `RecognizeEntitiesActionResult`, `RecognizeLinkedEntitiesActionResult`, and `RecognizePiiEntitiesActionResult`.
+
 ## 5.1.0-beta.7 (2021-05-19)
 ### Features Added
 - Added property `disableServiceLogs` to all endpoints' options bag